table.py: fix crash when a region overflows with remaining complete chunks
This patch fixes a bug identified by @jfrkuska in Issue #4.

The crashes are caused by a memory map file satisfying all of the
following conditions:

  1. Region "A" causes overflow dispatch to a next-level table;

  2. Region "A" has remaining complete chunks to map after that
     overflow dispatch, and those remaining complete chunks are
     permitted to be mapped as blocks in the current table i.e.
     do not need to be dispatched to a next-level table;

  3. The first chunk of Region "B" is a dispatch to the same
     next-level table dispatched to in point (1) above.

The crashes are caused by the overflow logic in table.py [1], which
does not reduce `num_chunks` by 1 to correctly account for the overflow
dispatch in point (1).

This leads the complete chunk mapping logic [2] to insert a leaf node
Region object in the current table's list of entries in point (2).

This confuses the dispatch in point (3), which expects to either:

  A) Find an empty entry, in which case it allocates a Table object
     and dispatches to that newly allocated Table, or;

  B) Find an existing entry, which it assumes is a Table object and
     attempts to dispatch to it.

Unfortunately, in (B) above it does find an existing entry but it is
the leaf node Region object inserted by the complete chunk mapping
logic. The Region object does not have a map() function to dispatch
to, leading to a crash when we try to invoke that missing function.

To fix the bug, simply correctly decrement `num_chunks` by 1 in the
overflow logic.

[1] https://github.com/ashwio/arm64-pgtable-tool/blob/master/pgtt/table.py#L130
[2] https://github.com/ashwio/arm64-pgtable-tool/blob/master/pgtt/table.py#L153